File renames and split up compose-test
