+++ src/test/test.ts
@@ -1,9 +1,13 @@
 import { test } from "./util/cappucino";
 
-import * as ComposeAlphaTest from "./compose-alpha-test"
+import * as ComposeAlphaTest from "./compose-test"
 import * as ExampleFileTest from "./example-file-test"
+import * as SchemaTest from "./schema-test"
+import * as WorkflowsTest from "./workflows-test"
 
 ComposeAlphaTest
 ExampleFileTest
+SchemaTest
+WorkflowsTest
 
 test();
